+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## WIP
 
+- Update `WebTransportClientConfig` to use `WebServerDestination`, which allows connecting to a WebTransport server via URL (useful when your server has certs for a domain name).
 - Update `h3` dependencies for the WebTransport server in the `renet2` crate to depend on the `h3-v0.0.4` tag.
 - Implement https://github.com/lucaspoffo/renet/pull/158.
 - Bump `bevy_replicon_renet2` to v0.0.4 for `bevy_replicon` v0.26.